refactor: split argparser
